+++ src/openeo_aggregator/backend.py
@@ -1,6 +1,7 @@
 import contextlib
 import datetime
 import functools
+import itertools
 import logging
 import time
 from collections import defaultdict
@@ -11,7 +12,7 @@ import flask
 import openeo_driver.util.view_helpers
 from openeo.capabilities import ComparableVersion
 from openeo.rest import OpenEoApiError, OpenEoRestError, OpenEoClientException
-from openeo.util import dict_no_none, TimingLogger, deep_get
+from openeo.util import dict_no_none, TimingLogger, deep_get, rfc3339
 from openeo_aggregator.config import AggregatorConfig, STREAM_CHUNK_SIZE_DEFAULT, CACHE_TTL_DEFAULT, \
     CONNECTION_TIMEOUT_RESULT, CONNECTION_TIMEOUT_JOB_START
 from openeo_aggregator.connection import MultiBackendConnection, BackendConnection, streaming_flask_response
@@ -106,12 +107,39 @@ class AggregatorCollectionCatalog(AbstractCollectionCatalog):
             else:
                 _log.info(f"Merging {cid!r} collection metadata from backends {by_backend.keys()}")
                 metadata = self._merge_collection_metadata(by_backend)
+            metadata = self._normalize_metadata(metadata)
             collections_metadata.append(metadata)
             internal_data.set_backends_for_collection(cid, by_backend.keys())
 
         return collections_metadata, internal_data
 
-    def _merge_collection_metadata(self, by_backend: Dict[str, dict]) -> dict:
+    @classmethod
+    def _normalize_metadata(cls, metadata: dict) -> dict:
+        cid = metadata.get("id", None)
+        if cid is None:
+            raise OpenEOApiException("Missing collection id in metadata")
+        if "links" not in metadata:
+            metadata["links"] = []
+        metadata["links"] = [l for l in metadata["links"] if l.get("rel") not in ("self", "parent", "root")]
+        if flask.has_app_context():
+            metadata["links"].append({
+                "href": flask.url_for("openeo.collections", _external=True),
+                "rel": "root"
+            })
+            metadata["links"].append({
+                "href": flask.url_for("openeo.collections", _external=True),
+                "rel": "parent"
+            })
+            metadata["links"].append({
+                "href": flask.url_for("openeo.collection_by_id", collection_id=cid, _external=True),
+                "rel": "self"
+            })
+        else:
+            _log.warning("Unable to provide root/parent/self links in collection metadata outside flask app context")
+        return metadata
+
+    @classmethod
+    def _merge_collection_metadata(cls, by_backend: Dict[str, dict], report=_log.warning) -> dict:
         """
         Merge collection metadata dicts from multiple backends
         """
@@ -123,54 +151,113 @@ class AggregatorCollectionCatalog(AbstractCollectionCatalog):
         cid = ids.pop()
         _log.info(f"Merging collection metadata for {cid!r}")
 
+        # Start with some initial/required fields
         result = {
             "id": cid,
+            "stac_version": max(list(getter.get("stac_version")) + ["0.9.0"]),
+            "title": getter.first("title", default=cid),
+            "description": getter.first("description", default=cid),
+            "type": getter.first("type", default="Collection"),
+            "links": [l for l in list(getter.concat("links")) if l.get("rel") not in ("self", "parent", "root")],
         }
 
-        result["stac_version"] = max(list(getter.get("stac_version")) + ["0.9.0"])
-        stac_extensions = sorted(getter.union("stac_extensions", skip_duplicates=True))
-        if stac_extensions:
-            result["stac_extensions"] = stac_extensions
-        # TODO: better merging of title and description?
-        result["title"] = getter.first("title", default=result["id"])
-        result["description"] = getter.first("description", default=result["id"])
-        keywords = getter.union("keywords", skip_duplicates=True)
-        if keywords:
-            result["keywords"] = keywords
+        # Generic field merging
+        # Notes:
+        # - `crs` is required by OGC API: https://docs.opengeospatial.org/is/18-058/18-058.html#_crs_identifier_list
+        # - `sci:doi` and related are defined at https://github.com/stac-extensions/scientific
+        for field in getter.available_keys(["stac_extensions", "keywords", "providers", "sci:publications"]):
+            result[field] = getter.concat(field, skip_duplicates=True)
+        for field in getter.available_keys(["deprecated"]):
+            result[field] = all(getter.get(field))
+        for field in getter.available_keys(["crs", "sci:citation", "sci:doi"]):
+            result[field] = getter.first(field)
+
+        # Summary merging
+        result["summaries"] = {}
+        summaries_getter = getter.select("summaries")
+        for summary_name in summaries_getter.keys():
+            if summary_name in [
+                "constellation", "platform", "instruments",
+            ]:
+                result["summaries"][summary_name] = summaries_getter.concat(summary_name, skip_duplicates=True)
+            elif summary_name.startswith("sar:") or summary_name.startswith("sat:"):
+                result["summaries"][summary_name] = summaries_getter.concat(summary_name, skip_duplicates=True)
+            else:
+                report(f"Unhandled merging of summary {summary_name!r}")
+
+        # Assets
+        if getter.has_key("assets"):
+            result["assets"] = {k: getter.select("assets").first(k) for k in getter.select("assets").keys()}
+
+        # All keys with special merge handling.
         versions = set(getter.get("version"))
         if versions:
-            # TODO: smarter version maximum?
+            # TODO: smarter version maximum? Low priority, versions key is not used in most backends.
             result["version"] = max(versions)
-        deprecateds = list(getter.get("deprecated"))
-        if deprecateds:
-            result["deprecated"] = all(deprecateds)
         licenses = set(getter.get("license"))
         result["license"] = licenses.pop() if len(licenses) == 1 else ("various" if licenses else "proprietary")
-        # TODO: .. "various" if multiple licenses apply ... links to the license texts SHOULD be added,
-        providers = getter.union("providers", skip_duplicates=True)
-        if providers:
-            result["providers"] = list(providers)
+
         result["extent"] = {
             "spatial": {
-                "bbox": getter.select("extent").select("spatial").union("bbox", skip_duplicates=True) \
+                "bbox": getter.select("extent").select("spatial").concat("bbox", skip_duplicates=True) \
                         or [[-180, -90, 180, 90]],
             },
             "temporal": {
-                "interval": getter.select("extent").select("temporal").union("interval", skip_duplicates=True) \
+                "interval": getter.select("extent").select("temporal").concat("interval", skip_duplicates=True) \
                             or [[None, None]],
             },
         }
-        result["links"] = list(getter.union("links"))
-        # TODO: combine cube:dimensions smarter?
-        cube_dimensions = getter.first("cube:dimensions")
-        if cube_dimensions:
-            result["cube:dimensions"] = cube_dimensions
-        # TODO merge existing summaries?
-        result["summaries"] = {
-            # TODO: use a more robust/user friendly backend pointer than backend id (which is internal implementation detail)
-            self.STAC_PROPERTY_PROVIDER_BACKEND: list(by_backend.keys())
-        }
-        # TODO: assets ?
+
+        if getter.has_key("cube:dimensions"):
+            cube_dim_getter = getter.select("cube:dimensions")
+            result["cube:dimensions"] = {}
+
+            # Spatial dimensions
+            for dim in cube_dim_getter.available_keys(["x", "y"]):
+                result["cube:dimensions"][dim] = cube_dim_getter.first(dim)
+                # TODO: check consistency of step and reference_system?
+                try:
+                    bounds = cube_dim_getter.select(dim).concat("extent")
+                    result["cube:dimensions"][dim]["extent"] = [min(bounds), max(bounds)]
+                except Exception as e:
+                    report(f"Failed to merge cube:dimensions.{dim}.extent: {e!r}")
+            # Temporal dimension
+            for dim in cube_dim_getter.available_keys(["t"]):
+                result["cube:dimensions"][dim] = cube_dim_getter.first(dim)
+                # TODO: check consistency of step?
+                try:
+                    t_starts = [e[0] for e in cube_dim_getter.select(dim).get("extent") if e[0]]
+                    t_ends = [e[1] for e in cube_dim_getter.select(dim).get("extent") if e[1]]
+                    result["cube:dimensions"][dim]["extent"] = [
+                        min(rfc3339.normalize(t) for t in t_starts) if t_starts else None,
+                        max(rfc3339.normalize(t) for t in t_ends) if t_ends else None
+                    ]
+                except Exception as e:
+                    report(f"Failed to merge cube:dimensions.{dim}.extent: {e!r}")
+
+            for dim in cube_dim_getter.available_keys(["bands"]):
+                result["cube:dimensions"][dim] = cube_dim_getter.first(dim)
+                try:
+                    # Find common prefix of bands
+                    # TODO: better approach? e.g. keep everything and rewrite process graphs on the fly?
+                    bands_iterator = cube_dim_getter.select(dim).get("values")
+                    prefix = next(bands_iterator)
+                    for bands in bands_iterator:
+                        prefix = [t[0] for t in itertools.takewhile(lambda t: t[0] == t[1], zip(prefix, bands))]
+                        if bands != prefix:
+                            report(f"Trimming bands {bands} to common prefix {prefix}")
+                    result["cube:dimensions"][dim]["values"] = prefix
+                except Exception as e:
+                    report(f"Failed to merge cube:dimensions.{dim}.extent: {e!r}")
+
+        # TODO: use a more robust/user friendly backend pointer than backend id (which is internal implementation detail)
+        result["summaries"][cls.STAC_PROPERTY_PROVIDER_BACKEND] = list(by_backend.keys())
+
+        ## Log warnings for improper metadata.
+        # license => Log warning for collections without license links.
+        license_links = [l for l in getter.concat("links") if l.get("rel") == "license"]
+        if result["license"] in ["various", "proprietary"] and not license_links:
+            _log.warning(f"Missing license links for collection: {cid}")
 
         return result
 
@@ -251,11 +338,11 @@ class AggregatorCollectionCatalog(AbstractCollectionCatalog):
         if len(by_backend) == 0:
             raise CollectionNotFoundException(collection_id=collection_id)
         elif len(by_backend) == 1:
-            # TODO: also go through _merge_collection_metadata procedure (for clean up/normalization)?
-            return by_backend.popitem()[1]
+            metadata = by_backend.popitem()[1]
         else:
             _log.info(f"Merging metadata for collection {collection_id}.")
-            return self._merge_collection_metadata(by_backend=by_backend)
+            metadata = self._merge_collection_metadata(by_backend=by_backend)
+        return self._normalize_metadata(metadata)
 
     def load_collection(self, collection_id: str, load_params: LoadParameters, env: EvalEnv) -> DriverDataCube:
         raise RuntimeError("openeo-aggregator does not implement concrete collection loading")