+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated (generous fallback) "FreeTier" user role to 30DayTrial (more strict)
 - Use EODC dev instance in aggregator dev config
 - Update EGI issuer URL to new Keycloak one (keep old provider under "egi-legacy")
+- Improve collection metadata merging ([#5](https://github.com/Open-EO/openeo-aggregator/issues/5))
 
 ### Fixed